Fixing the broken image build + optimizations
The previous image build was broken - I think llvm was moved to GitHub.
Changed the build to clone the repo from the new URL, with a few more steps for clang and polly.

There were a lot of steps which made the cache quite big - merged some steps so that they don't eat up the cache.

Also added a few optimizations, like removing some tmp files, and cleaning the repo once llvm is installed. This reduced the image size from 58 GB to  24 GB.